+++ app/src/main/java/ca/etsmtl/applets/seemobile/model/Poste.java
@@ -3,6 +3,8 @@ package ca.etsmtl.applets.seemobile.model;
 import com.j256.ormlite.field.DatabaseField;
 import com.j256.ormlite.table.DatabaseTable;
 
+import java.util.Date;
+
 /**
  * Created by gnut3ll4 on 22/12/15.
  */
@@ -33,8 +35,22 @@ public class Poste {
     @DatabaseField
     private boolean aPostule;
 
-    @DatabaseField(foreign = true, foreignAutoCreate = true, foreignAutoRefresh = true)
-    private PosteDetails details;
+    @DatabaseField
+    private String codePostal;
+    @DatabaseField
+    private String coordonnateur;
+    @DatabaseField
+    private Date datePostulationLimite;
+    @DatabaseField
+    private String description;
+    @DatabaseField
+    private String langueCv;
+    @DatabaseField
+    private String missionEmployeur;
+    @DatabaseField
+    private String siteWebEmployeur;
+    @DatabaseField
+    private String typeEmployeur;
 
     public Poste() {
     }
@@ -103,11 +119,67 @@ public class Poste {
         this.aPostule = aPostule;
     }
 
-    public PosteDetails getDetails() {
-        return details;
+    public String getCodePostal() {
+        return codePostal;
     }
 
-    public void setDetails(PosteDetails details) {
-        this.details = details;
+    public void setCodePostal(String codePostal) {
+        this.codePostal = codePostal;
+    }
+
+    public String getCoordonnateur() {
+        return coordonnateur;
+    }
+
+    public void setCoordonnateur(String coordonnateur) {
+        this.coordonnateur = coordonnateur;
+    }
+
+    public Date getDatePostulationLimite() {
+        return datePostulationLimite;
+    }
+
+    public void setDatePostulationLimite(Date datePostulationLimite) {
+        this.datePostulationLimite = datePostulationLimite;
+    }
+
+    public String getDescription() {
+        return description;
+    }
+
+    public void setDescription(String description) {
+        this.description = description;
+    }
+
+    public String getLangueCv() {
+        return langueCv;
+    }
+
+    public void setLangueCv(String langueCv) {
+        this.langueCv = langueCv;
+    }
+
+    public String getMissionEmployeur() {
+        return missionEmployeur;
+    }
+
+    public void setMissionEmployeur(String missionEmployeur) {
+        this.missionEmployeur = missionEmployeur;
+    }
+
+    public String getSiteWebEmployeur() {
+        return siteWebEmployeur;
+    }
+
+    public void setSiteWebEmployeur(String siteWebEmployeur) {
+        this.siteWebEmployeur = siteWebEmployeur;
+    }
+
+    public String getTypeEmployeur() {
+        return typeEmployeur;
+    }
+
+    public void setTypeEmployeur(String typeEmployeur) {
+        this.typeEmployeur = typeEmployeur;
     }
 }
